# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenClaw Context Engine API
  description: |
    OpenClaw 上下文引擎中间层 API —— 为 AI Agent 提供持久记忆和结构化知识检索。
    
    ## 架构
    - **localMem**: SQLite 记忆存储（2-state 模型：tentative / kept）
    - **LLMWiki**: 结构化知识页面编译和搜索
    
    ## 认证
    配置 `OPENCLAW_API_SECRET` 后，所有请求需要 Bearer Token。
    Unix Socket 连接自动注入 Token。
    
    ## 速率限制
    默认 100 请求/分钟，可通过环境变量配置。
  version: 1.0.0
  contact:
    name: OpenClaw Team
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:8901
    description: 本地 HTTP 服务
  - url: http://localhost
    description: Unix Socket 代理（通过 /tmp/openclaw-engine.sock）
security:
  - bearerAuth: []
...
paths:
  /api/memory/query:
    post:
      tags: [Memory]
      summary: 查询记忆
      description: 搜索记忆条目，支持可选的 Wiki 结果合并
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryQueryRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryQueryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          description: 查询仅包含元数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataOnlyError'
        '429':
          $ref: '#/components/responses/RateLimit'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/memory/query-context:
    post:
      tags: [Memory]
      summary: 查询记忆上下文
      description: 查询记忆并返回关联的会话和轮次信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryQueryContextRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryQueryContextResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/memory/save:
    post:
      tags: [Memory]
      summary: 保存记忆
      description: 保存一条记忆，可选走治理流程
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemorySaveRequest'
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemorySaveResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/memory/{id}:
    get:
      tags: [Memory]
      summary: 获取单条记忆
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Memory]
      summary: 更新记忆内容
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 5000
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Memory]
      summary: 删除记忆
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  memory_id:
                    type: string
                  action:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/memory/turn:
    post:
      tags: [Memory]
      summary: 追加对话轮次
      description: 追加一轮对话到当前会话，触发异步 autoTriage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionTurnRequest'
      responses:
        '200':
          description: 追加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/memory/session/start:
    post:
      tags: [Memory]
      summary: 启动记忆会话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartSessionRequest'
      responses:
        '200':
          description: 会话启动成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  session_id:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/memory/reviews:
    get:
      tags: [Memory]
      summary: 列出待审核记忆
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewsResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/memory/reviews/{id}/promote:
    post:
      tags: [Memory]
      summary: 提升待审核记忆为永久
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      responses:
        '200':
          description: 提升成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewActionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/memory/reviews/{id}/discard:
    post:
      tags: [Memory]
      summary: 丢弃待审核记忆
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      responses:
        '200':
          description: 丢弃成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewActionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/memory/reviews/{id}/evaluate:
    post:
      tags: [Memory]
      summary: 评估待审核记忆
      description: 存储 LLM 评估结果，不修改记忆状态
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                evaluation:
                  type: object
                  properties:
                    score:
                      type: number
                    reasoning:
                      type: string
                    recommendation:
                      type: string
      responses:
        '200':
          description: 评估成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewActionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/memory/auto-triage:
    post:
      tags: [Memory]
      summary: 单条 autoTriage
      description: 对一条 turn 执行自动沉淀判断
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, role, content]
              properties:
                session_id:
                  type: string
                role:
                  type: string
                  enum: [user, assistant, system]
                content:
                  type: string
                previous_role:
                  type: string
                previous_content:
                  type: string
      responses:
        '200':
          description: triage 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  candidates:
                    type: array
                  count:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/memory/auto-triage/batch:
    post:
      tags: [Memory]
      summary: 批量 autoTriage
      description: 处理指定时间范围内的 turns
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                hours:
                  type: integer
                  default: 24
                  minimum: 1
                  maximum: 168
      responses:
        '200':
          description: 批量 triage 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  processed:
                    type: integer
                  extracted:
                    type: integer
                  daily_limit_reached:
                    type: boolean
        '500':
          $ref: '#/components/responses/InternalError'
  /api/memory/session/reset:
    post:
      tags: [Memory]
      summary: 重置当前活跃会话
      responses:
        '200':
          description: 重置成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '500':
          $ref: '#/components/responses/InternalError'
  /api/memory/session/import-transcript:
    post:
      tags: [Memory]
      summary: 导入外部对话记录为会话
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                transcript_path:
                  type: string
                transcript_id:
                  type: string
                transcripts_root:
                  type: string
                project_id:
                  type: string
                title:
                  type: string
                created_at:
                  type: string
                  format: date-time
                session_id:
                  type: string
      responses:
        '200':
          description: 导入成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  session_id:
                    type: string
                  imported_turn_count:
                    type: integer
                  status:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/memory/governance/plan-update:
    post:
      tags: [Memory]
      summary: 治理流程预演
      description: 返回更新计划但不实际写入
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                aliases:
                  type: array
                  items:
                    type: string
                path_hints:
                  type: array
                  items:
                    type: string
                collection_hints:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: 规划成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  strategy:
                    type: string
                  suggestedMemoryId:
                    type: string
                  relatedMemoryIds:
                    type: array
                    items:
                      type: string
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API 认证密钥（OPENCLAW_API_SECRET）
  parameters:
    MemoryId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: 记忆 ID
  schemas:
    MemoryQueryRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
          description: 查询文本
        top_k:
          type: integer
          default: 3
          minimum: 1
          maximum: 100
        include_wiki:
          type: boolean
          default: false
          description: 是否合并 Wiki 搜索结果
    MemoryQueryContextRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
        top_k:
          type: integer
          default: 3
        session_id:
          type: string
        include_debug:
          type: boolean
          default: false
...
    MemoryQueryResponse:
      type: object
      properties:
        query:
          type: string
        hits:
          type: array
          items:
            $ref: '#/components/schemas/MemoryHit'
        total:
          type: integer
        freshness_level:
          type: string
    MemoryQueryContextResponse:
      allOf:
        - $ref: '#/components/schemas/MemoryQueryResponse'
        - type: object
          properties:
            matched_sessions:
              type: array
            matched_turns:
              type: array
            context:
              type: object
...
    MemoryResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/MemoryItem'
    MemorySaveResponse:
      type: object
      properties:
        success:
          type: boolean
        memory_id:
          type: string
        content:
          type: string
        state:
          type: string
        status:
          type: string
...
    MetadataOnlyError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: query_contains_only_metadata
        message:
          type: string
        reason:
          type: string
...
  responses:
    BadRequest:
      description: 请求验证失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: VALIDATION_ERROR
            message: "Validation failed: query is required"
...
    RateLimit:
      description: 速率限制
      headers:
        Retry-After:
          schema:
            type: integer
          description: 重试等待秒数
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: RATE_LIMIT_EXCEEDED
            message: "Rate limit exceeded. Try again in 60 seconds."
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: INTERNAL_ERROR
            message: "An unexpected error occurred"
...